Custom gapps dev folder
There are roms that use their own gapps packages and store them in
goo.im

As it's not possible to handle these packages the same way as the
official ones (there's a spacific url to retrieve if there's a new
official gapps version), i've added the option to check a custom folder
in goo to retrieve the last package found in it.

Note that in the selected folder there must be only gapps packages.